# contracts/contract3-api.openapi.yaml
openapi: 3.1.0
info:
  title: DarkPulse Investigator API
  version: 1.0.0
paths:
  /auth/login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
  /auth/me:
    get:
      tags:
      - Auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
...
  /alerts/config:
    get:
      tags:
      - Alerts
      summary: Get Alert Config
      operationId: get_alert_config_api_v1_alerts_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
    put:
      tags:
      - Alerts
      summary: Update Alert Config
      operationId: update_alert_config_api_v1_alerts_config_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertConfig'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /alerts/history:
    get:
      tags:
      - Alerts
      summary: Get Alert History
      operationId: get_alert_history_api_v1_alerts_history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /alerts/history/{alert_id}:
    patch:
      tags:
      - Alerts
      summary: Patch Alert
      operationId: patch_alert_api_v1_alerts_history__alert_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          title: Alert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /alerts/ws-ticket:
    post:
      tags:
      - Alerts
      summary: Ws Ticket
      operationId: ws_ticket_api_v1_alerts_ws_ticket_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      security:
      - HTTPBearer: []
...
  /alerts/ws:
    get:
      tags:
      - Alerts
      summary: AlertsWebsocket
      description: Live alert stream. Authenticate with a single-use ticket from POST
        /auth/ws-ticket passed as ?ticket=.
      parameters:
      - name: ticket
        in: query
        required: true
        schema:
          type: string
      responses:
        '101':
          description: Switching Protocols
components:
  schemas:
    AlertConfig:
      properties:
        rules:
          items:
            $ref: '#/components/schemas/AlertRule'
          type: array
          title: Rules
      type: object
      title: AlertConfig
    AlertHistoryResponse:
      properties:
        data:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: AlertHistoryResponse
    AlertPatch:
      properties:
        acknowledged:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Acknowledged
        assignee:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Assignee
        resolved:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Resolved
      type: object
      title: AlertPatch
    AlertRule:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        severity_min:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Severity Min
          default: 0
        products:
          items:
            type: string
          type: array
          maxItems: 200
          title: Products
        neighborhoods:
          items:
            type: string
          type: array
          maxItems: 200
          title: Neighborhoods
        enabled:
          type: boolean
          title: Enabled
          default: true
        audience:
          type: string
          enum:
          - viewer
          - analyst
          - administrator
          title: Audience
          default: viewer
      type: object
      required:
      - name
      title: AlertRule
    ApiEnvelope:
      properties:
        data:
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: ApiEnvelope
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - code
      - message
      title: ErrorDetail
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        token:
          type: string
          maxLength: 4096
          minLength: 1
          title: Token
      type: object
      required:
      - token
      title: LoginRequest
    Pagination:
      properties:
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        limit:
          type: integer
          title: Limit
          default: 50
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: Pagination
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8003